[Mod] change to use PySide6 instead of PyQt5
